Update README.md
LINKED LIST
Linked list is a linear data structure in which elements are not stored in contiguous memory locations. We can insert and delete a node at any point of linked list. It contains nodes where each node contains data field and a reference link. We do not have to initialize the size of it just like in array.